// src/api/models/Card.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Card = {
    id?: number;
    cardName?: string;
    creditLimit?: number;
    currentBalance?: number;
    deleted?: boolean;
    createdAt?: string;
    lastModifiedAt?: string;
};